# Improved ObjectId parsing and ToString() performance